Standardize `Window.onDropFile` in JS.
Now instead of getting a `FileList` object, you get a string URL just
like on other platforms. If the file is an image, this URL can be loaded
via `Image.loadFromFile()`, as expected. If not, it can be loaded using
any other standard way to load URLs.

Since loading is required either way, I don't believe this hurts
performance in any meaningful way. However, it does allocate an object
URL that won't be deallocated until the page is closed or the user calls
`js.html.URL.revokeObjectURL()`. I'm not too worried about this.